refactor: move mock data
