+++ src/interfaces/response.ts
@@ -26,53 +26,9 @@ interface Photo {
   likes: number;
   liked_by_user: boolean;
   current_user_collections: never[];
-  sponsorship: Sponsorship | null;
+  sponsorship: object | null;
   topic_submissions: object;
   user: User;
-}
-
-interface Sponsorship {
-  impression_urls: never[];
-  tagline: string;
-  tagline_url: string;
-  sponsor: {
-    id: string;
-    updated_at: string;
-    username: string;
-    name: string;
-    first_name: string;
-    last_name: null;
-    twitter_username: string;
-    portfolio_url: string;
-    bio: string;
-    location: string;
-    links: {
-      self: string;
-      html: string;
-      photos: string;
-      likes: string;
-      portfolio: string;
-      following: string;
-      followers: string;
-    };
-    profile_image: {
-      small: string;
-      medium: string;
-      large: string;
-    };
-    instagram_username: string;
-    total_collections: number;
-    total_likes: number;
-    total_photos: number;
-    accepted_tos: boolean;
-    for_hire: boolean;
-    social: {
-      instagram_username: string;
-      portfolio_url: string;
-      twitter_username: string;
-      paypal_email: null;
-    };
-  };
 }
 
 interface User {
@@ -82,104 +38,12 @@ interface User {
   name: string;
   first_name: string;
   last_name: string | null;
-  twitter_username: string | null;
-  portfolio_url: string | null;
-  bio: string | null;
-  location: string | null;
-  links: {
-    self: string;
-    html: string;
-    photos: string;
-    likes: string;
-    portfolio: string;
-    following: string;
-    followers: string;
-  };
-  profile_image: {
-    small: string;
-    medium: string;
-    large: string;
-  };
-  instagram_username: string | null;
-  total_collections: number;
-  total_likes: number;
-  total_photos: number;
-  accepted_tos: boolean;
-  for_hire: boolean;
-  social: {
-    instagram_username: string | null;
-    portfolio_url: string | null;
-    twitter_username: string | null;
-    paypal_email: string | null;
-  };
-}
-
-interface Tag {
-  type: string;
-  title: string;
-  source?: {
-    ancestry: {
-      type: Ancestry;
-      category?: Ancestry;
-      subcategory?: Ancestry;
-    };
-    title: string;
-    subtitle: string;
-    description: string;
-    meta_title: string;
-    meta_description: string;
-    cover_photo: CoverPhoto;
-  };
-}
-
-interface Ancestry {
-  slug: string;
-  pretty_slug: string;
-}
-
-interface CoverPhoto {
-  id: string;
-  created_at: string;
-  updated_at: string;
-  promoted_at: string | null;
-  width: number;
-  height: number;
-  color: string;
-  blur_hash: string;
-  description: string | null;
-  alt_description: string | null;
-  urls: {
-    raw: string;
-    full: string;
-    regular: string;
-    small: string;
-    thumb: string;
-    small_s3: string;
-  };
-  links: {
-    self: string;
-    html: string;
-    download: string;
-    download_location: string;
-  };
-  likes: number;
-  liked_by_user: boolean;
-  current_user_collections: [];
-  sponsorship: Sponsorship | null;
-  topic_submissions: object;
-  premium?: boolean;
-  plus?: boolean;
-  user: User;
-}
-
-interface PhotoWithTags extends Photo {
-  tags: Tag[];
 }
 
 interface SearchResult {
   total: number;
   total_pages: number;
-  results: PhotoWithTags[];
+  results: Photo[];
 }
 
-export type { Photo, Sponsorship, User, SearchResult };
+export type { Photo, User, SearchResult };
